requirements.txt
pyPdf==1.13
beautifulsoup4==4.3.2
Markdown==2.6.5
requests==2.8.1
PySocks==1.5.6

